r/aws_ssoadmin_account_assignment: Move functions around.
